[C++] Moved fcl related structures to multibody/fcl.hpp + rework somes inclusion directives
