+++ src/parsers/urdf.hpp
@@ -23,9 +23,6 @@
 #include "pinocchio/deprecated.hh"
 #ifdef WITH_HPP_FCL
   #include "pinocchio/multibody/geometry.hpp"
-  #include <hpp/fcl/collision_object.h>
-  #include <hpp/fcl/collision.h>
-  #include <hpp/fcl/shape/geometric_shapes.h>
 #endif
 
 #include <urdf_model/model.h>
